% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{The ERA5 annual seasonal maximum Fire Weather Index and Monthly Summary Record of Fire Weather Index for the region of New South Wales from 1978 to 2019}
\format{
\subsection{'data'}{

A data frame with 43 rows and 3 columns
\describe{
\item{year}{the year of the value}
\item{FWI}{The annual seasonal maximum Fire Weather Index}
\item{MSR}{Monthly Summary Record of Fire Weather Index}
}
}
}
\source{
\url{https://climexp.knmi.nl/bushfires_timeseries.cgi}
}
\usage{
data
}
\description{
The ERA5 annual seasonal maximum Fire Weather Index and Monthly Summary Record of Fire Weather Index for the region of New South Wales from 1978 to 2019
}
\keyword{datasets}
