% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fwi_data}
\alias{fwi_data}
\title{Annual ERA5 averages of soil moisture for the region of New South Wales from 1978 to 2019}
\format{
\subsection{'fwi_data'}{

A data frame with 43 rows and 2 columns
\describe{
\item{year}{the year of the value}
\item{FWI}{Annual average Fire Weather Index}
}
}
}
\source{
\url{https://climexp.knmi.nl/bushfires_timeseries.cgi}
}
\usage{
fwi_data
}
\description{
Annual ERA5 averages of soil moisture for the region of New South Wales from 1978 to 2019
}
\keyword{datasets}
