% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_fwi_dashboard}
\alias{run_fwi_dashboard}
\title{Run the FWI dashboard}
\usage{
run_fwi_dashboard()
}
\value{
None
}
\description{
Run the FWI dashboard
}
\examples{
if (interactive()){
run_fwi_dashboard()
}
}
